refactor: extract horizontal-layout core CSS to styles folder
