+++ packages/horizontal-layout/src/vaadin-horizontal-layout.js
@@ -9,8 +9,8 @@ import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { horizontalLayoutStyles } from './styles/vaadin-horizontal-layout-styles.js';
 import { HorizontalLayoutMixin } from './vaadin-horizontal-layout-mixin.js';
-import { horizontalLayoutStyles } from './vaadin-horizontal-layout-styles.js';
 
 /**
  * `<vaadin-horizontal-layout>` provides a simple way to horizontally align your HTML elements.
